Extracted Bids3D Dataset
